requirements.txt
cryptography==2.6.1
Flask==1.0.2
pycrypto==2.6.1
pyTelegramBotAPI==3.6.6